+++ src/main/java/com/alechilles/alecstamework/commands/TameworkCommandRoot.java
@@ -30,6 +30,7 @@ public final class TameworkCommandRoot extends AbstractCommandCollection {
         addSubCommand(new TameworkDebugHookCommand());
         addSubCommand(new TameworkDebugSpawnerCommand());
         addSubCommand(new TameworkDebugPromptCommand());
+        addSubCommand(new TameworkDebugDespawnCommand());
         addSubCommand(new TameworkDebugLagCommand());
     }
 }